test: añadir cobertura al 100% en new code de SonarCloud
- AnonymousProjectDetailResponse, AsignarEvaluadoresRequest, EvaluarEvaluacionRequest
- EvaluacionService: consultarDetalleAnonimo, asignarEvaluadores, evaluar
- GetResolutionUseCase: findByProcedureId
- ResolutionController: getByProcedureId
- ResolutionRepositoryAdapter: findByProcedureId
- ProcedureRecentActivity: getters/setters
- AuditLogRepository + AuditLogController
- 38 evaluacion tests total, 1247 tests total, BUILD SUCCESS

+++ fiis/src/main/java/com/sgi/fiis/grupos_investigacion/infrastructure/persistence/ResearchGroupJpaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResearchGroupJpaRepository extends JpaRepository<ResearchGroupEntity, Integer> {
 
-    ResearchGroupEntity findByCurrentCoordinatorId(Integer coordinatorId);
+    ResearchGroupEntity findByCurrentCoordinatorId(Long coordinatorId);
 }
