+++ fiis/src/main/java/com/sgi/fiis/tramites/infrastructure/persistence/SpringDataProcedureRepository.java
@@ -18,7 +18,7 @@ public interface SpringDataProcedureRepository extends JpaRepository<ProcedureEn
 
     List<ProcedureEntity> findByStatusAndReviewerRole(String status, String reviewerRole);
 
-    List<ProcedureEntity> findByStatusAndGroupId(String status, Integer groupId);
+    List<ProcedureEntity> findByStatusAndGroupId(String status, Long groupId);
 
     boolean existsByCode(String code);
 }
